Update google-calendar extension
- fix: event dates and categories
- Merge branch \'master\' of https://github.com/d4vz/google-calendar
- feat: add extension icon
- Update README.md
- Merge branch \'master\' of https://github.com/d4vz/google-calendar
- fix: remove unused images
- Update README.md
- Update README.md
- Update README.md
- feat: update readme
- fix: remove logout
- feat: add icon in extension manifest
- fix: type error
- feat: use autorize in fetch
- Update README.md
- feat: add step by step to configure
- Merge branch \'master\' of https://github.com/d4vz/google-calendar
- fix: use env
- feat: add readme description
