+++ extensions/google-calendar/.gitignore
@@ -8,6 +8,7 @@ raycast-env.d.ts
 .raycast-swift-build
 .swiftpm
 compiled_raycast_swift
+.env
 
 # misc
 .DS_Store
