+++ extensions/google-calendar/src/services/calendar.ts
@@ -1,17 +1,9 @@
-import fetch from "node-fetch";
-
-import { client } from "../client";
+import { fetch } from "../lib/fetch";
 import { Calendar } from "../types/calendar";
 
 export async function getUserCalendars() {
   const url = "https://www.googleapis.com/calendar/v3/users/me/calendarList";
-  const tokens = await client.getTokens();
-
-  const headers = {
-    Authorization: `Bearer ${tokens?.accessToken}`,
-  };
-
-  const response = await fetch(url, { headers });
+  const response = await fetch(url);
 
   if (!response.ok) {
     console.error("fetch calendars error:", await response.text());
